third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/content-element-select-namespace-expected.txt

CONSOLE ERROR: line 9: Uncaught TypeError: h.createShadowRoot is not a function
